feat(STOCK-20) minor refactoring, switched models to dtos
